Convert ClassicOption implementation to TypeScript
