+++ packages/option-t/src/classic_option/classic_option.ts
@@ -1,4 +1,4 @@
-import type { TransformFn, RecoveryFn, EffectFn } from '../internal/function.ts';
+import type { TransformFn, RecoveryFn, EffectFn } from '../internal/function.js';
 
 /**
  *  @deprecated
@@ -175,28 +175,220 @@ export abstract class ClassicOptionBase<T> implements ClassicOptionable<T> {
     private readonly ok: boolean;
     private readonly val: T | undefined;
 
-    readonly isSome: boolean;
-    readonly isNone: boolean;
+    protected constructor(ok: boolean, val: T | undefined) {
+        this.ok = ok;
+        this.val = val;
+        Object.seal(this);
+    }
 
-    protected constructor(ok: boolean, val: T | undefined);
+    /**
+     *  Return whether this is `Some<T>` or not.
+     */
+    get isSome(): boolean {
+        return this.ok;
+    }
 
-    unwrap(): T | never;
-    unwrapOr(def: T): T;
-    unwrapOrElse(fn: RecoveryFn<T>): T;
-    expect(msg: string): T | never;
-    map<U>(fn: TransformFn<T, U>): ClassicOption<U>;
-    flatMap<U>(fn: ClassicTryTransformFn<T, U>): ClassicOption<U>;
-    mapOr<U>(def: U, fn: TransformFn<T, U>): U;
-    mapOrElse<U>(def: RecoveryFn<U>, fn: TransformFn<T, U>): U;
-    and<U>(optb: ClassicOption<U>): ClassicOption<U>;
-    andThen<U>(fn: ClassicTryTransformFn<T, U>): ClassicOption<U>;
-    or(optb: ClassicOption<T>): ClassicOption<T>;
-    orElse(fn: ClassicTryRecoveryFn<T>): ClassicOption<T>;
-    drop(destructor?: EffectFn<T>): void;
+    /**
+     *  Return whether this is `None` or not.
+     */
+    get isNone(): boolean {
+        return !this.ok;
+    }
+
+    /**
+     *  Returns the inner `T` of a `Some<T>`.
+     *  @throws {TypeError}
+     *      Throws if the self value equals `None`.
+     */
+    unwrap(): T | never {
+        if (!this.ok) {
+            throw new TypeError('called `unwrap()` on a `None` value');
+        }
+
+        // This code is required to keep backward compatibility.
+        // @ts-expect-error
+        return this.val;
+    }
+
+    /**
+     *  Returns the contained value or a default value `def`.
+     */
+    unwrapOr(def: T): T {
+        // This code is required to keep backward compatibility.
+        // @ts-expect-error
+        return this.ok ? this.val : def;
+    }
+
+    /**
+     *  Returns the contained value or computes it from a closure `fn`.
+     */
+    unwrapOrElse(fn: RecoveryFn<T>): T {
+        // This code is required to keep backward compatibility.
+        // @ts-expect-error
+        return this.ok ? this.val : fn();
+    }
+
+    /**
+     *  Returns the inner `T` of a `Some<T>`.
+     *  @throws {TypeError}
+     *      Throws a custom error with provided `msg`
+     *      if the self value equals `None`.
+     */
+    expect(msg: string): T | never {
+        if (!this.ok) {
+            throw new TypeError(msg);
+        }
+
+        // This code is required to keep backward compatibility.
+        // @ts-expect-error
+        return this.val;
+    }
+
+    /**
+     *  Maps an `Option<T>` to `Option<U>` by applying a function to a contained value.
+     */
+    map<U>(fn: TransformFn<T, U>): ClassicOption<U> {
+        if (!this.ok) {
+            // cheat to escape from a needless allocation.
+            // @ts-expect-error
+            return this;
+        }
+
+        // This code is required to keep backward compatibility.
+        // @ts-expect-error
+        const value = fn(this.val);
+        const option = createClassicSome(value);
+        return option;
+    }
+
+    /**
+     *  Returns `None` if the self is `None`,
+     *  otherwise calls `fn` with the wrapped value and returns the result.
+     */
+    flatMap<U>(fn: ClassicTryTransformFn<T, U>): ClassicOption<U> {
+        if (!this.ok) {
+            // cheat to escape from a needless allocation.
+            // @ts-expect-error
+            return this;
+        }
+
+        // This code is required to keep backward compatibility.
+        // @ts-expect-error
+        const mapped = fn(this.val);
+        const isOption = mapped instanceof ClassicOptionBase;
+        if (!isOption) {
+            throw new TypeError("Option<T>.flatMap()' param `fn` should return `Option<T>`.");
+        }
+
+        return mapped;
+    }
+
+    /**
+     *  Applies a function `fn` to the contained value or returns a default `def`.
+     */
+    mapOr<U>(def: U, fn: TransformFn<T, U>): U {
+        if (this.ok) {
+            // This code is required to keep backward compatibility.
+            // @ts-expect-error
+            return fn(this.val);
+        } else {
+            return def;
+        }
+    }
+
+    /**
+     *  Applies a function `fn` to the contained value or computes a default result by `defFn`.
+     */
+    mapOrElse<U>(def: RecoveryFn<U>, fn: TransformFn<T, U>): U {
+        if (this.ok) {
+            // This code is required to keep backward compatibility.
+            // @ts-expect-error
+            return fn(this.val);
+        } else {
+            return def();
+        }
+    }
+
+    /**
+     *  Returns `None` if the self is `None`, otherwise returns `optb`.
+     */
+    and<U>(optb: ClassicOption<U>): ClassicOption<U> {
+        // This code is required to keep backward compatibility.
+        // @ts-expect-error
+        return this.ok ? optb : this;
+    }
+
+    /**
+     *  The alias of `Option<T>.flatMap()`.
+     */
+    andThen<U>(fn: ClassicTryTransformFn<T, U>): ClassicOption<U> {
+        return this.flatMap(fn);
+    }
+
+    /**
+     *  Returns the self if it contains a value, otherwise returns `optb`.
+     *
+     *  @template   T
+     *
+     *  @param  {!Option<T>} optb
+     *  @return {!Option<T>}
+     */
+    or(optb: ClassicOption<T>): ClassicOption<T> {
+        // This code is required to keep backward compatibility.
+        // @ts-expect-error
+        return this.ok ? this : optb;
+    }
+
+    /**
+     *  Returns the self if it contains a value,
+     *  otherwise calls `fn` and returns the result.
+     */
+    orElse(fn: ClassicTryRecoveryFn<T>): ClassicOption<T> {
+        if (this.ok) {
+            // This code is required to keep backward compatibility.
+            // @ts-expect-error
+            return this;
+        } else {
+            const value = fn();
+            if (value instanceof ClassicOptionBase) {
+                return value;
+            }
+
+            throw new TypeError("Option<T>.orElse()' param `fn` should return `Option<T>`.");
+        }
+    }
+
+    /**
+     *  Finalize the self.
+     *  After this is called, the object's behavior is not defined.
+     *
+     *  @param destructor
+     *      This would be called with the inner value if self is `Some<T>`.
+     */
+    drop(destructor?: EffectFn<T>): void {
+        if (this.ok && typeof destructor === 'function') {
+            // This code is required to keep backward compatibility.
+            // @ts-expect-error
+            destructor(this.val);
+        }
+
+        // This code is required to keep backward compatibility.
+        // @ts-expect-error
+        this.val = null;
+        Object.freeze(this);
+    }
+
     // FYI: this is json representation.
     // eslint-disable-next-line @typescript-eslint/naming-convention
-    toJSON(): { is_some: boolean; value: T | undefined };
+    toJSON(): { is_some: boolean; value: T | undefined } {
+        return {
+            // eslint-disable-next-line @typescript-eslint/naming-convention
+            is_some: this.ok,
+            value: this.val,
+        };
+    }
 }
+Object.freeze(ClassicOptionBase.prototype);
 
 /**
  *  @deprecated
@@ -253,24 +445,43 @@ export type ClassicOption<T> = ClassicSome<T> | ClassicNone<T>;
  *  @deprecated
  *      See https://github.com/option-t/option-t/issues/459
  */
-export declare function createClassicSome<T>(val: T): ClassicSome<T>;
+export function createClassicSome<T>(val: T): ClassicSome<T> {
+    // This code is required to keep backward compatibility.
+    // @ts-expect-error
+    const o = new ClassicOptionBase(true, val);
+    return o;
+}
 
 /**
  *  @deprecated
  *      See https://github.com/option-t/option-t/issues/459
  */
-export declare function createClassicNone<T>(): ClassicNone<T>;
+export function createClassicNone<T>(): ClassicNone<T> {
+    // This code is required to keep backward compatibility.
+    // @ts-expect-error
+    const o = new ClassicOptionBase(false, undefined);
+    return o;
+}
 
 /**
  *  @deprecated
  *      See https://github.com/option-t/option-t/issues/459
  */
 // eslint-disable-next-line @typescript-eslint/naming-convention
-export declare const ClassicSomeConstructor: new <T>(val: T) => ClassicSome<T>;
+export const ClassicSomeConstructor: new <T>(val: T) => ClassicSome<T> =
+    function ClassicSomeConstructor<T>(val: T) {
+        return createClassicSome(val);
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    } as any;
 
 /**
  *  @deprecated
  *      See https://github.com/option-t/option-t/issues/459
  */
 // eslint-disable-next-line @typescript-eslint/naming-convention
-export declare const ClassicNoneConstructor: new <T>() => ClassicNone<T>;
+export const ClassicNoneConstructor: new <T>() => ClassicNone<T> = function ClassicNoneConstructor<
+    T
+>() {
+    return createClassicNone<T>();
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+} as any;
